Split shell.rs into "runners"
The initial runners are `docker` and `native`, but more could be added
later allowing for extensions like we already have for storage
`Backend`s.

No attempt to make a `Runner` trait yet, this commit just splits the
code in `shell.rs`.
