+++ src/runners/mod.rs
@@ -1,0 +1,7 @@
+pub use self::{
+    docker::{docker_run, DockerRunFlags, ShellModes},
+    native::native_run,
+};
+
+mod docker;
+mod native;
